requirements.txt
Flask==3.0.3
boto3==1.34.148
gunicorn==22.0.0
python-dotenv==1.0.1